# ARROW-10785: [Rust] Optimize take string