--- a/rust/arrow/benches/take_kernels.rs
+++ b/rust/arrow/benches/take_kernels.rs
@@ -47,19 +47,20 @@ where
     Arc::new(array) as ArrayRef
 }
 
-fn create_strings(size: usize) -> ArrayRef {
-    let v = (0..size)
-        .map(|_| {
-            seedable_rng()
-                .sample_iter(&Alphanumeric)
-                .take(5)
-                .collect::<String>()
-        })
-        .collect::<Vec<_>>();
+fn create_strings(size: usize, null_density: f32) -> ArrayRef {
+    let rng = &mut seedable_rng();
 
-    Arc::new(StringArray::from(
-        v.iter().map(|x| &**x).collect::<Vec<&str>>(),
-    ))
+    let mut builder = StringBuilder::new(size);
+    for _ in 0..size {
+        let x = rng.gen::<f32>();
+        if x < null_density {
+            let value = rng.sample_iter(&Alphanumeric).take(4).collect::<String>();
+            builder.append_value(&value).unwrap();
+        } else {
+            builder.append_null().unwrap()
+        }
+    }
+    Arc::new(builder.finish())
 }
 
 fn create_random_index(size: usize, null_density: f32) -> UInt32Array {
@@ -122,27 +123,41 @@ fn add_benchmark(c: &mut Criterion) {
         b.iter(|| bench_take(&values, &indices))
     });
 
-    let values = create_strings(512);
+    let values = create_strings(512, 0.0);
     let indices = create_random_index(512, 0.0);
     c.bench_function("take str 512", |b| b.iter(|| bench_take(&values, &indices)));
 
-    let values = create_strings(1024);
+    let values = create_strings(1024, 0.0);
     let indices = create_random_index(1024, 0.0);
     c.bench_function("take str 1024", |b| {
         b.iter(|| bench_take(&values, &indices))
     });
 
-    let values = create_strings(512);
+    let values = create_strings(512, 0.0);
     let indices = create_random_index(512, 0.5);
-    c.bench_function("take str nulls 512", |b| {
+    c.bench_function("take str null indices 512", |b| {
         b.iter(|| bench_take(&values, &indices))
     });
 
-    let values = create_strings(1024);
+    let values = create_strings(1024, 0.0);
     let indices = create_random_index(1024, 0.5);
-    c.bench_function("take str nulls 1024", |b| {
+    c.bench_function("take str null indices 1024", |b| {
         b.iter(|| bench_take(&values, &indices))
     });
+
+    let values = create_strings(1024, 0.5);
+
+    let indices = create_random_index(1024, 0.0);
+    c.bench_function("take str null values 1024", |b| {
+        b.iter(|| bench_take(&values, &indices))
+    });
+
+    let values = create_strings(1024, 0.5);
+    let indices = create_random_index(1024, 0.5);
+    c.bench_function("take str null values null indices 1024", |b| {
+        b.iter(|| bench_take(&values, &indices))
+    });
+
 }
 
 criterion_group!(benches, add_benchmark);