Further optimizes take string to benefit from creating a buffer directly.

Interestingly, it seems it is still faster to copy a `Vec`  than to use a buffer for the the string values. Presumably `extend_from_slice` is much slower than the `Vec` one.

This PR:

* pre allocates buffer for offsets
* adds a special case for 0 null indices in which case we can skip a check
* adding some benchmarks for string value arrays containing nulls

Benchmark results

```
take str 512            time:   [2.1942 us 2.1956 us 2.1971 us]                          
                        change: [-22.950% -22.847% -22.738%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low severe
  1 (1.00%) high mild
  1 (1.00%) high severe

take str 1024           time:   [3.7476 us 3.7529 us 3.7605 us]                           
                        change: [-23.958% -23.825% -23.678%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe

take str null indices 512                                                                             
                        time:   [2.0983 us 2.1015 us 2.1056 us]
                        change: [-24.742% -24.387% -23.981%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe

take str null indices 1024                                                                             
                        time:   [3.6767 us 3.6788 us 3.6812 us]
                        change: [-22.588% -22.497% -22.406%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

take str null values 1024                                                                             
                        time:   [3.6732 us 3.6773 us 3.6836 us]
                        change: [-23.029% -22.857% -22.695%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

take str null values null indices 1024                                                                             
                        time:   [5.3789 us 5.3821 us 5.3853 us]
                        change: [-10.711% -10.610% -10.509%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
```